refactor: info 페이지 파트별로 분리
