+++ src/components/my/info/commonStyle.module.scss
@@ -1,0 +1,57 @@
+.title {
+  @include pretendard-32-500;
+
+  color: $grey-900;
+  display: flex;
+  align-items: center;
+  gap: 15px;
+
+  & > .subTitle {
+    @include pretendard-20-500;
+
+    color: red;
+  }
+}
+
+.publicCheckboxWrapper {
+  display: flex;
+  gap: 10px;
+  align-items: center;
+
+  & > input {
+    width: 20px;
+    height: 20px;
+    border: 1px solid $grey-500;
+    cursor: pointer;
+  }
+
+  & > span {
+    @include pretendard-16-400;
+
+    color: $grey-500;
+  }
+}
+
+.divider {
+  width: 1px;
+  min-height: 45px;
+  height: 100%;
+  background-color: $grey-500;
+  margin: 0 45px;
+}
+
+.editInputWrapper {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+}
+
+.addButton {
+  @include pretendard-20-500;
+
+  padding: 5px 15px;
+  width: max-content;
+  border-radius: 5px;
+  cursor: pointer;
+  margin-left: auto;
+}
