+++ src/components/my/info/introducePart.module.scss
@@ -1,0 +1,20 @@
+@import './commonStyle.module';
+
+.introduceWrapper {
+  @include pretendard-24-400;
+
+  width: 100%;
+  height: 150px;
+  color: black;
+  padding: 25px 20px;
+  border: 1px solid $grey-500;
+  border-radius: 5px;
+  background-color: $grey-50;
+  margin-top: 15px;
+  overflow-y: scroll;
+  resize: none;
+
+  &:focus {
+    outline: none;
+  }
+}
