+++ src/components/my/info/myProfileCard.module.scss
