+++ src/components/my/info/specializationPart.module.scss
@@ -1,0 +1,17 @@
+@import './commonStyle.module';
+
+.speciallizationWrapper > * {
+  @include pretendard-24-400;
+
+  color: $grey-700;
+}
+
+.speciallizationDropdown {
+  @include pretendard-24-400;
+
+  padding: 18px 20px;
+  color: $grey-700;
+  width: 480px;
+  height: 65px;
+  margin-top: 30px;
+}
