--- src/components/my/skillInput.module.scss
@@ -1,28 +1,0 @@
-.container {
-  display: flex;
-  height: 70px;
-  margin-bottom: 20px;
-}
-
-.skillNameInput {
-  @include pretendard-24-400;
-
-  color: $grey-500;
-  height: 100%;
-  flex-grow: 1;
-  background-color: $grey-50;
-  border-radius: 5px 0 0 5px;
-  border: 1px solid $grey-500;
-  padding: 18px 20px;
-
-  &:focus {
-    outline: none;
-  }
-}
-
-.levelCheckboxWrapper {
-  padding: 18px 20px;
-  border-radius: 0 5px 5px 0;
-  border: 1px solid $grey-500;
-  border-left: none;
-}
