+++ src/pages/my-page/info/info.module.scss
@@ -4,180 +4,3 @@
   gap: 90px;
   margin-top: 45px;
 }
-
-.speciallizationWrapper > * {
-  @include pretendard-24-400;
-
-  color: $grey-700;
-}
-
-.speciallizationDropdown {
-  @include pretendard-24-400;
-
-  padding: 18px 20px;
-  color: $grey-700;
-  width: 480px;
-  height: 65px;
-  margin-top: 30px;
-}
-
-.title {
-  @include pretendard-32-500;
-
-  color: $grey-900;
-  display: flex;
-  align-items: center;
-  gap: 15px;
-
-  & > .subTitle {
-    @include pretendard-20-500;
-
-    color: red;
-  }
-}
-
-.publicCheckboxWrapper {
-  display: flex;
-  gap: 10px;
-  align-items: center;
-
-  & > input {
-    width: 20px;
-    height: 20px;
-    border: 1px solid $grey-500;
-    cursor: pointer;
-  }
-
-  & > span {
-    @include pretendard-16-400;
-
-    color: $grey-500;
-  }
-}
-
-.manageText {
-  @include pretendard-20-500;
-
-  color: $grey-500;
-  cursor: pointer;
-}
-
-.introduceWrapper {
-  @include pretendard-24-400;
-
-  width: 100%;
-  height: 150px;
-  color: black;
-  padding: 25px 20px;
-  border: 1px solid $grey-500;
-  border-radius: 5px;
-  background-color: $grey-50;
-  margin-top: 15px;
-  overflow-y: scroll;
-  resize: none;
-}
-
-.individualInfoWrapper {
-  margin-top: 30px;
-  margin-bottom: 30px;
-  display: flex;
-  flex-direction: column;
-}
-
-.infoInput {
-  margin-top: 15px;
-  margin-bottom: 28px;
-}
-
-.editInputWrapper {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-}
-
-.addButton {
-  @include pretendard-20-500;
-
-  padding: 5px 15px;
-  width: max-content;
-  border-radius: 5px;
-  cursor: pointer;
-  margin-left: auto;
-}
-
-.individualInfoWrapper > .list {
-  @include pretendard-28-500;
-
-  color: $grey-800;
-  letter-spacing: -0.28px;
-  display: flex;
-
-  & > .label {
-    min-width: 75px;
-  }
-}
-
-.skillInfoWrapper {
-  margin-top: 30px;
-  margin-bottom: 30px;
-  width: 100%;
-  height: max-content;
-}
-
-.skillInfoWrapper > .list {
-  width: 100%;
-  height: auto;
-  display: flex;
-  align-items: center;
-  letter-spacing: -0.28px;
-  line-height: 45px;
-
-  & > .label {
-    @include pretendard-28-500;
-
-    min-width: 75px;
-  }
-
-  & > .content {
-    @include pretendard-24-400;
-
-    width: 425px;
-    height: max-content;
-  }
-}
-
-.experienceWrapper {
-  @include pretendard-28-400;
-
-  margin-top: 30px;
-  display: flex;
-  flex-direction: column;
-}
-
-.experienceWrapper > .list {
-  @include pretendard-24-400;
-
-  display: flex;
-  align-items: center;
-  line-height: 45px;
-
-  & > .content {
-    width: 374px;
-  }
-
-  & > .date {
-    width: 236px;
-  }
-}
-
-.portfolioWrapper {
-  width: 700px;
-}
-
-.divider {
-  width: 1px;
-  min-height: 45px;
-  height: 100%;
-  background-color: $grey-500;
-  margin: 0 45px;
-}
